Update NPM dependencies
This update exclusively relates to client-side libraries distributed as part of the application. The update to DashJS is just a minor patch with a couple of bug fixes. I don't actually believe we're using Headroom currently, but it's a minor update. The GSAP update includes a major version; apparently, this library was significantly out of date. That said, all animations continue to work as expected.