Removed legacy transformation for debug mode
In previous versions of ASP.NET, the debug state was configured in the `web.config` using the `<compilation debug="true" />` attribute. That is no longer used in ASP.NET Core, but the transformation to remove it still persisted in our `Web.Release.config` transformation. Removing this doesn't change any behavior, but removes a warning we received during publishing a _Release_ profile.