Set `hostingModel` to lower case value
Per the documentation, "In-process hosting is set with InProcess, which is the default value. The value of <AspNetCoreHostingModel> is case insensitive, so inprocess and outofprocess are valid values." While that may be, however, Visual Studio complains that the value doesn't match the enumeration constraint. To quiet this warning, we're accepting the lowercase value.