Set `processPath`
The `processPath` was already set in the `web.config`—and that should be the only location it's actually required. Visual Studio, however, throws a warning due to the required attribute missing in the `Web.Release.config`—even though this is _exclusively_ used to set the scope of the subsequent transformation. As such, it _shouldn't_ be required here, and has no impact on the actual application itself.