Removed `Microsoft.Data.SqlClient` dependency
The `Microsoft.Data.SqlClient` is now implicitly included as part of the `OnTopic.Data.Sql` implementation of the `ITopicRepository` and, thus, shouldn't need to be explicitly referenced as part of the project. This was previously needed to ensure that Azure DevOps Pipelines could publish a working version to Azure App Services with all dependencies.