Fix reference to `IsEditorComponent()`
Breaking changes should not be introduced between major versions when using Semantic Versioning. This rule was inadvertantly broken with the introduction of OnTopic Editor 4.1.0, which moved `IsEditorComponent()` to a static method per the recommendations of Code Analysis. While this may well be permitted, and even advisable, I missed that this was a publicly exposed method. Fortunately, it's an easy change, even if it's an unfortunate mistake.