Introduced server-side validation for invoices
In a previous commit, we introduced invoice validation based on invoice number and amount. This was enabled using the `[Remote()]` validation attribute on the `PaymentFormBindingModel`. Unfortunately, it appears that Microsoft doesn't re-execute that validation on the server and, therefore, it's _conceivable_ for a customer to get to the server with an invalid invoice (amount) if e.g. they're able to submit the form prior to e.g. the web service confirming that the invoice is invalid on the client. To mitigate this, I've duplicated the validation on the server. This _shouldn't_ be necessary, but also isn't _that_ much code.