Rolling back Foundation Sites
Upgrading to 6.6.3 introduced an issue with the distribution of the navigation menu's links. To mitigate that, we're revering back to 6.6.1 until we're able to better understand—and hopefully resolve!—the issue.

Note that we tested this in 6.6.2 as well, and that appears to be when the issue was introduced. So this isn't new to 6.6.3, that's just when we first noticed it.