Convert invoice amounts to doubles
The original version of invoice validation used the `GetInteger()` method to retrieve the `InvoiceAmount` from the topic graph. This works fine assuming that the value can be converted to an `int`. For values with decimal points, however, this resulted in rounding or truncating the value—and, thus, failing validation when the customer put in the correct value. This is fixed by retrieving them as a `string` (since we don't have a `GetDouble()` method) and then converting them to a `double` using `Double.TryParse()`.