Introduced basic `LessonListViewComponent`
The `LessonListViewComponent` will list all siblings of the current page in what will be exposed via a left-hand navigation.

This doesn't yet support progress tracking; that will be added in a future implementation.

A smarter version of this might ascend to a `Unit` content type and display its children—similar to the `PageLevelNavigationViewComponent`—but as we don't expect lessons to be hierarchical, this isn't currently required.