Implemented basic progress tracking for Lessons
With the help of a new `IsVisited()` method, this code will now a) write a cookie for the current page if it doesn't already exist, and b) set the `IsVisited` property for each `TrackedNavigationTopicViewModel` in the `NavigationRoot` based on that cookie value. This allows us to track the progress of users based on simple cookie-based storage, without the need for authentication.