Implemented basic progress tracking for Units
In the `LessonListViewComponent`, a cookie is written for each unit based on the total number of lessons completed. This cookie value is then evaluated in the `UnitListViewComponent`, using an `IsComplete()` helper method, to conditionally set the `IsVisited` property of the `TrackedNavigationTopicViewModel`.

The `IsVisited` property is nullable specifically for this use case, thus supporting three states: Not started (`null`), started but incomplete (`false`), and complete (`true`).