Introduced `CoursesController` and corresponding route
The `CoursesController` is basically just a `TopicController`, except it adds one piece of functionality to the `Index()` method: It will redirect a `Unit` content type to the first child. This allows the course to be broken down by logical units without there needing to be a unit page/template. This is very similar to how the `PageGroup` content type is handled in the **OnTopic Library**.

As part of this, the `/Courses` route is established, and the `CoursesController` is added to the `GoldSimActivator`.