Setup dependency injection of new view components
The new `UnitListViewComponent` and `LessonListViewComponent` must be wired up in the _Componsition Root_ in order to have dependencies injected. Because they depend on a different `INavigationTopicViewModel` than the existing view components, they also require the creation of a new `IHierarchicalTopicMappingService<>`.

This will allow these view components to be used in actual views.