Introduced `UnitTopicViewModel`
The `UnitTopicViewModel` exposes the `UnitNumber` property. Ideally, this would be automatically generated on the client, but that's not easy to establish in all cases; for that reason, this is instead treated as a mapped attribute. The `UnitTopicViewModel` is then referenced as a parent attribute on `LessonTopicViewModel`, giving it strongly typed access to the unit data.