Exposed progress tracking via Font-Awesome
Added a `GetCssClass()` method on `TrackedNavigationTopicViewModel` and use it to inject a Font Awesome glyph for progress tracking on both the `LessonList` as well as the `UnitList`. For now, basic style support is hard-coded into `Lesson.cshtml`; this will need to be centralized into the SCSS later.

In addition, updated `LessonListViewComponent` to utilize the nullability of `IsVisited`. This isn't strictly necessary since a lesson is either complete or unstarted; we don't track whether it's partially complete. But by unifying the state tracking with units, it allows us to use the same `GetCssClass()` for both units and lessons by treating `null` as unstarted, `false` as incomplete, and `true` as complete.