Move progress tracking CSS into `_Lists.scss`
This allows us to centralize the CSS definitions, take advantage of SCSS, and ensures that the output is both minimized and cached.

As part of this, renamed the class to `track-progress` so that it's a bit more generalized, and not specific to e.g. the `UnitList` or `LessonList` components (even though that's the only place we're _currently_ using them.)