Implemented semantic version of Font Awesome
Removed Font Awesome markup and explicit classes from views, instead integrating Font Awesome directly into the semantic CSS (e.g., `unstarted`, `incomplete`, `complete`). There are some limitations to this—namely the inability to work with anything outside of the _Regular_ icons—but it has the benefit of being more semantic, reducing the markup, and further centralizing control of the styling.