Explicitly set the sort order of example applications
By default, relationships are sorted by the `Topic.Id`—which correlates to how they're indexed in the database. For technical reasons, it's difficult to retain their original order. In fact, it may even be entirely meaningless, since it would require attempting to determine the appropriate order between items in different trees.

Given this, if order is relevant to how relationships are displayed, they should be explicitly ordered by an available attribute. (The default of `Topic.Id` roughly correlates to the date created—but even that can be misleading in some circumstances.)

To mitigate this for example applications listed on the application page template, the order is explicitly set to reverse chronological order based on the last modified date, thus showing the most recently _updated_ topics at the top.