Conditionally add Razor runtime compilation for development mode
Razor runtime compilation allows views to be dynamically compiled upon file change, similar to how they behaved in ASP.NET Core 2.x and ASP.NET MVC 5.x.

This adds overhead to the first request for each view, as it must be dynamically compiled. As such, while this is _useful_ to add to the _development_ environment, thus allowing real-time edits to views, it isn't preferrable for the _production_ environment. To mitigate this, it is conditionally enabled based on the current environment.