Remove redundant `grid-container`
The Foundation `grid-container` effectively resets the grid coordinates for the layout, thus allowing e.g. a new grid to start anew off of a 12-column grid without needing to be aware of or worry about what any ascendent grids might be doing. This is useful. In the `_Layout.cshtml`, however, the parent `<article />` element is _already_ assigned `grid-container`, and so there isn't any purpose or benefit to _resetting_ the grid within the `PageBodySection`.

This has been tested on both pages and forms—such as the `/Pay` form—including pages with tables, and partial-width callout boxes to ensure that it doesn't have any impact on current layouts. Not every page was tested, but the sampling of pages evaluated should be representative.