Removed extraneous `div`
On the `Page` template, there is an extaneous `div` which doesn't appear to serve a purpose. There aren't any apparent selectors which rely on this for formatting—nor ought there be, as that's a messy way of handling selections. Perhaps at some point this had a class assigned to it which has since been removed, and thus it's vestigal.

Regardless, removing it doesn't _appear_ to impact the layout, so we're removing it.