Removed `.body` class on `Page` view
The `.body` class is a bit ambiguous semantically, but it is generally used to control margins around ``.pane` and `.embedded` elements. The one place where it's used in conjunction with `.body` is in support for _baubles_ a piece of illustrative decor that we implemented, but never rolled out. That said, even for baubles, that selector can easily be modified not to _depend_ upon `.body`.

Removing it doesn't have any apparent impact on the page template. And since this is a self-contained template, it should be pretty easy to identify off-target effects. There doesn't appear to be any selectors that use it, for instance, to manage HTML content embedded via the `@Model.Body`.