Removed extraneous `article` from multi-part templates
Several templates—namely `ApplicationPage`, `TechnicalPaperList`, and `Video`—each compile the contents of their pages from multiple attributes. For instance, the `ApplicationPage`'s main `panel` contains the `Abstract`, `Body`, and then various lists of e.g. `ApplicationExample`s.

In these cases, to distinguish the main body area, it was wrapped in an `<article class="page body" />`. Presumably, this was to semantically describe the main content area for e.g. screenreaders—but it didn't include schema.org properties such as `itemscope`, `itemtype`, and `itemprop`, which continued to be applied to the parent `<article />`. As such, this was at best ambiguous, and more likely useless in terms of providing semantic value.

Second, because of a selector on `article.page.body`, these sections have an additional 3rem bottom margin. That probably makes sense in general for the main `<article />`, but it's debatable in context of these composite page templates. For instance, on `ApplicationPage`, it creates a bigger margin between the body and subsequent header than any of the other sections and their subsequent headers. It looks fine, but it's inconsistent, and likely an off-target affect that was not intentional.

It's worth noting that the one place where this extraneous margin might make good sense is on the video page, where it creates some buffer between the body and the video. If we want to maintain this, however, we could easily add a margin to the `video` element or the `.video-wrapper` class.

For this reason, I'm removing these duplicate `<article />` declarations. If, in the future, we want to revisit defining these as the primary schema.org article, that's fine—but that's a bigger project, as it necessitates removing the parent `<article />` declarations.