Remove `.page` and `.content` from `main`
While the `page` and `content` classes may have had an original semantic purpose in their application to the `<main />` element, they haven't been applied in any meaningful way. There isn't any definition in the SCSS for `.content`, and the only definitions for `.page`—outside of the page-level navigation—relate exclusively to `#MainContentArea article.page.body` selector.

Given this, I'm removing both `.page` and `.content` from the `<main />` element. If we find we need some semantic class at this level in the future, we can revisit this. But that should be a more deliberate effort with a careful understanding of the purpose and meaning of the class name.