Picked up vagrant `embedded pane`
In 0517ab4, we collapse `section.embedded.pane.body` to `header .body`. It appears I missed one in `Accordion.cshtml`, however. This remedies that. See 0517ab4 for further explanation of this change. In short, however, it just simplifies the selector and begins standardizing the nomenclature a bit more by removing extraneous classes which correlate heavily with one another.