Refactored `h1.title.short` as `header.page-banner h1.title`
Originally, a general selector was established for `<h1 class="title short" />`. This was only used in two places on the site, however: in the Page-Level Navigation, and for the Homepage. There were two problems with this. First, most of the styles used for the Page-Level Navigation weren't relevant to the home page—and, in fact, were being overwritten by `Home.scss`. Second, neither of these titles are actually _short_, so the semantic meaning isn't clear.

Since most of the styles used by the Page-Level Navigation are specific to the Page-Level Navigation, I've updated the selector to use `header.page-banner h1.title`. (`header.page-banner` is the container for the Page-Level Navigation). This allows us to remove the `short` class from the Page-Level Navigation.

As part of that, I also copied the `margin: 0` setting to the `Home.scss` definition, since that's the _one_ attribute which appears to be used by both.

This makes for a more intuitive selector, and effectively eliminates the `short` class name outside of the homepage. (That will need to be addressed in a future update.)