Renamed `short` and `long` to `kicker` and `headline` on home page
The `short` and `long` titles on the home page weren't particularly descriptive, in that both were long, and they mapped to the `Title` and `Subtitle` attribtues, not the `ShortTitle` and `Title` attributes, as one might expect. Further, `short` and `long` have been used elsehwere on the site, but with different meanings (namely, the main `Title` is `long`, and the Page-Level Navigation was `short`).

Since these styles are really specific to the home page, and already disagree in usage from the names they inherited from the `Page` content type in OnTopic, it makes more sense to establish new names that help indicate their purpose. For this, I've borrowed from newspaper nomenclature to call them `kicker` and `headline` respectively. (In a newspaper, the _kicker_ is the small text above the headline of an article which provides context.)

We can revisit these names later should we choose. But, for now, they help disambiguiate the `short` and `long` nomenclature on the site, while also making it more semantic. The fact that these aren't used elsewhere also makes it really clear that these are, in fact, home-page specific—which makes sense, given that this type of setup doesn't really exist anywhere else.

While I was at it, I also removed the `page` class, which had no selectors assocated with it—at least for `h1`s—and had otherwise been removed elsewhere on the site as part of a previous commit (a5acb8f).