Removed `long` title; now it's just `title`
Previously, the main title on the site needed to have both `title` and `long` classes; `title` gave it its size, and `long` gave it its color. But, outside of the Page-Level Navigation, `title` and `long` are always used together—and exclusively for the main title of the page. This means they're tightly coupled in a way that doesn't add much value semantically, and causes confusion as to the distinction.

In c9ec8bd, we addressed the first part of this, by specifically targeting `header.page-banner h1.title` for the Page-Level Navigation, to distinguish it from the main title. In this update, we remove `long` entirely, relying exclusively on `h1.title` and `article header h1.title` for the main title.

This doesn't appear to impact either the Page-Level Navigation nor the home page, both of which also rely on an `article header h1.title`; those remain compatible with the current styles, and override those that are specific to them. This is a bit of surprise since the Page-Level Navigation and the homepage's kicker previously used `short` instead of `long`. This probably gives them some additional styles they don't _need_, but it doesn't _hurt_ anything.