Removed `cell` from titles
The `cell` class is part of Foundation's grid structure. It's a useful class, and establishes an element as part of the grid, thus working in conjunction with other classes such as `medium-4` or `small-8` which determine how many grid columns an element should occupy.

That said, there isn't any real need or benefit for applying this to the title, which will always take up the entire width. Further, even if this was valuable, it would make more sense to apply at a higher level, such as the `<header />`, whereas right now it's been applied to `<h1 class="title" />` but not e.g. `<h2 class="subtitle" />` and other elements in the header.

For now, we're just removing it in order to clean up the markup and remove unnecessary classes. If we have a need to have the title—or header more generally-participate in the grid structure in a more meaningful way, then we can either reintroduce it at that point, or mix it in using SCSS for the appropriate selectors (e.g., `header` or `.title`).