Removed `body` from `Slideshow` template
The `.panel.body` class usually adds an extra margin to the bottom of a `panel`. In the case of the carousel, however, this is somewhat irrelevant since the inner `carousel-wrapper` dynamically copies and moves the `owl-nav` template above and below the `carousel-wrapper`, thus establishing its own margins that are greater than the `body`. As such, I've removed `body` since it provides no purpose or benefit.

As an aside, I _attempted_ to combine the `carousel-wrapper` into the `<section class="panel" />`. This works, but it results in the bottom margins being messed up. Namely, it pushes the buttons down, and resuts in them sliding underneath the footer, even when the calls to action are "docked". I'm not entirely sure _why_ that is, but am also not entirely _surprised_ since Owl Carousel has its own logic for handling the `carousel-wrapper`, and that may well compete with our `panel` styles which were not designed to account for it.

Regardless, while it might have been nice to have consolidate these, it's not a big deal to retain this extra level of nesting for the couple of places where we implement the Owl Carousel—namely on the `Slideshow` and `Home` templates.