Eliminated redundant `.accordion`
Previously, an `accordion` class was not only established on the parent `section`, but also on each nested `ul`. Since the `ul` is always nested under the `<section class="accordion" />`, however, it makes more sense to target it using e.g. `.accordion ul` instead of having a separate `section.accordion` and `ul.accordion`.

In order to support this, I needed to introduce the `accordion-container` mixin into the `.accordion > ul` selector, since that mapes to the element that we place `data-accordion` on. Among other things, this sets the `list-style-type` to `none`. It seems this looks for `ul.accordion` by default, so to remov this and avoid the duplicate class names, we needed to explicitly introduce that.