Applied `grid-x` to `PostTitleSection`, instead of entire `header`
The `cell` places the element into Foundation's grid struture—and, without further classes, instructs it to take up the width of all columns. I had previously removed this class from the `<h1 class="title" />` (167099e) with the idea that it shouldn't be necessary. But, without it, pages with a kicker (`category`)—such as the Example Application—end up showing the `h2` and `h1` on the same line.

There are two ways of mitigating this. The first is to remove the `grid-x`. But that prevents e.g. metadata in the `PostTitleSection` from utilizing the grid. The second is to add the `cell` to the `h1` (and, really, it should be on each of the headers). The third is to move the `grid-x` to only cover the metadata.

Since we shouldn't need the grid structure for the titles, I've chosen the third approach, thus limiting this exclusively to the place where we need it.