Reintroduced `body` into `Slideshow` template
In a recent commit (24d8a4f), I removed the `body` class from the `Slideshow` template as part of a general effort to remove this semantically ambiguous class from the templates (ab5cc50). The effect of the `body` class is to introduce a bottom margin to the element, and it is usually combined with the `panel` class to create that margin.

Unfortunately, the `body` class—or, at least, that bottom margin—is still necessary for the `Slideshow` template. Without it, the calls to action end up overlapping with the bottom paging buttons in certain cases.

In general, I suspect we would be better served by just defining a bottom margin on the `panel` class itself. This would also address an issue identified with the courseware project in which two back-to-back panels don't have any space between them. Until then, however, I'm reintroducing `body` back into the `Slideshow`.

Ultimately, we need to find a better solution (or simply identifier) for `body`, but for now it remains a necessity in multiple cases, even while we've been able to remove it in others.