Removed `body` from `panel` declarations
In a previous commit, we removed the `body` class from the `Page` template (ab5cc50), as the only sector targeting `<section class="body" />` is for the header (for cases where the body is places outside of a `panel`).

The `Page` template isn't the only scenario where `body` was used in conjunction with the `panel`, however. This update removed all of those unnecessary (vestigal?) declarations of `body` from e.g. `<section class="panel body" />`.

With this update, there are only two cases which remain:
1. Areas where the body is embedded in the `header`, outside of the `panel`,
2. Inline `article` declarations, where it's used to add a bottom margin,

The first scenario will be persisted, for now. The second scenario will be the subject of a future commit.