Resolved bottom margin issue for `Slideshow`
In a previous commit, `body` had been removed from `Slideshow` (24d8a4f). This caused an issue with the bottom margin overlapping with Owl Carousel's controls. As such, that change had been undone in a subsequent commit (f84df39).

This commit returns `body` _again_, effectively returning us back to the original commit (24d8a4f). In addition, however, it updates the selector in `Slideshow.scss` to only target `.panel` instead of `.panel.body`. Since this style sheet is exclusively called from `Slideshow.cshtml`, and there is only one `panel` on the Slideshow template, there's no reason to include the `body` class as part of the selector—and, thus, markup. While I was at it, I added `asp-append-version` to the stylesheet link, which makes it easier to test stylesheet changes like this.

Apologies for the messy commit history. Fortunately, this ultimately settled on a clean solution!