Removed `body` from `Search`
The `Search` view has an inline `article` with a `body` class. This class is responsible for adding a bottom margin. This is necessary for two reasons. First, the `simpleNavigation` plugin we use for paging places the navigation buttons outside of the content area, thus occupying the `panel`'s margin. Second, for reasons that are unclear to me, the `pagination` class was zeroeing out the bottom margin specifically.

To mitigate this, I changed the `pagination` class's `margin-bottom` to `rem-calc(32)`, which is the same value used, by default, for the `panel` class, thus recreating the `panel`'s bottom margin.

This isn't 100% identical, to note, as `panel` also has media queries for altering this for smaller screens. We're not going to fuss with that here, as we generally don't reproduce that level of fine-tuning on page-level classes. (Though, in the future, we may wish to centralize it via a mix-in in order to facilitate that.)

This allows us to remove `body`. While I was at it, I also added `asp-append-version` to the stylesheet to make these types of updates easier to evaluate, and to ensure end users end up with the correct updates, even if they have a cached version of the previous stylesheet.