Removed extraneous `article` from `Search`
In 7da5943, we removed an extraneous `article` from several mutli-part templates. Presumably, this had been added as a semantic hint to aid screen readers. In practice, this just introduced ambiguity since the view is wrapped in another `article` in `_Layout.cshtml`—and that `article` contains the schema.org annotations used by most e.g. screenreaders.

This update applies that same fix to `Search`. It had been skipped with 7da5943 because it included several other classes. The `search` and `results` classes, while potentially being semantically interesting, are not actually used in practice. If we want to revisit them later, we can add them to the parent `article` element. But for now they serve no benefit. In addition, the `article` had a `body` element, which _did_ have a style associated with it, but that was refactored to remove the `body` class in the previous commit (f653b2a).