Moved margin from `body` to new `.content.list` selector
In 7da5943, we removed an extraneous `article` from several mutli-part templates. Presumably, this had been added as a semantic hint to aid screen readers. In practice, this just introduced ambiguity since the view is wrapped in another `article` in `_Layout.cshtml`—and that `article` contains the schema.org annotations used by most e.g. screenreaders.

This update applies that same fix to both `LinkedList` and `ContentList`. Because the `body` class was being utilized here to add a bottom margin, that margin was instead moved to a new `.content.list` selector in `_lists.scss`.

(Note: This may not be the best location for it since it's not a `ul` or `ol`; if needed, we can move it later. There aren't enough distinct styles for these templates to justify a stand-alone view at this point, though.)

As part of this, I moved the `h1` to inside of the `<section class="content list" />`. This was already done on `ContentList`, but hadn't been done on `LinkedList` for some reason. If the `h1` remained on the outside, the selector would need to be applied to it instead, since we want the margin _above_ the header, not between the header and the articles. Semantically, it is preferred for headers to be _within_ the sections they describe, so I adopted that convention from `ContentList`.