Refactored sibling selector for embedded body
In a previous commit, the `.embedded.body` was moved to `header .body` (0517ab4). As part of that, a sibling selector was skipped as it wasn't clear what it was used for. It's still not clear what this was intended for, but it doesn't appear to be in use anywhere currently. Nevertheless, I'm reluctant to remove it without understanding its purpose.

Given that, since the `embedded` class has otherwise been removed, I've updated this sibling selector to simply look for `body`. Normally, when the body is placed outside of a `panel`, it is instead placed _within_ the `header`—and, thus, the new `header .body` selector. This allows the body to be placed _adjacent_ to the `header` instead. Most notably, this introduces significant padding on the left and right side.

This treatment is very similar to how the homepage treats out-of-panel content blocks, so this may have been useful for providing similar treatment on other views. Again, though, since it's not currently used, that's not clear. Given that, I've inserted a not to flag that it's not in use, so we can reevaluate it later.