Remove redundant `grid-container`
In 1d192cf, I removed the reduntant `grid-container` class around `@RenderBody()` in `_Layout.cshtml` because it was already nested within a parent element (`<article />`) which reset the Foundation grid. As such, any elements in e.g. `@Model.Body` which rely on that grid structure would be cordoned off from the broader grid structure used for the site layout.

This commit applies the same change to the `Payments.cshtml`. Since the `Payments.cshtml` is a form, and contains tight control over the layout in order to maintain multiple columns for form fields, it makes sense that it would want to operate in context of a fresh `grid-container`. As the `grid-container` was reset in the parent `<article />` as part of the `_Layout.cshtml`, however, there isn't any value to doing it within `Payments.cshtml`; they are _already_ firewalled from the overall page layout grid.

Likewise, there isn't any need to have a _separate_ `grid-container` for the body and the form, as they both implement their own `grid-x`. The twelve column grid from the parent `<article />` is satisfactory for both.