Applied refactorings from `maintenance/SCSS-Layout-Refactoring`
With `maintenance/SCSS-Layout-Refactoring` (60eb562), I made a number of refactorings to the markup and CSS in order to simplify the stylesheets, make them more semantic, and reduce unnecessary or redundant elements and styles.

As the courseware templates were created prior to these refactorings, they maintained some of the markup and styles affected by that refactoring. This commit updates the `Course` and `Lesson` template to apply those changes—which largely entails removing unnecessary markup and classes.

For instance, `grid-container` is already set on the parent of `PageBodySection`, so there's no benefit to "resetting" the grid immediately after. Also, the `body`, `page`, and `long` classes were depreecated in most scenarios. Finally, there's no need for the header to be in a `grid-x` unless it needs more sophisticated formatting—and, thus, we can also remove `cell.