Introduced `CourseListViewComponent`
The `CourseListViewComponent` is essentially identical to the `UnitListViewComponent`, except that it is intended to be used on the `/Courses` root. As both simply display their child topics, however, alongside progress tracking, the basic logic is identical—and, as such, `CourseListViewComponent` simply inherits from `UnitListViewComponent`. In fact, we could have simply reused `UnitListViewComponent` for this purpose, but having a separate wrapper is a) more semantically descriptive, and b) simplifies the ability to have a distinct view for presenting courses.

The latter is critical. Currently, the view is just a copy of the `UnitListViewComponent` template. In the future, however, we'll be modifying this to, likely, be a list of `panel`s, each fully linked, and including an abstract description—or something along those lines. In that way, while the business logic and data model may be similar, the presentation is expected to vary considerably.