Established `CourseList` content type, view model, and view
The `CourseList` is very similar to the `UnitList` content type, in that it is really just a basic page, and defers most of the logic to the `CourseListViewComponent` (53d830c). For now, the view is identical to `UnitList`—though, in the final version, we'll likely opt for an "embedded" body, and handle the `panel` as part of the `CourseListViewComponent` itself (as we'll end up with a clickable `panel` for each course). For now, this establishes the basic scaffolding to handle this.