Introduced progress tracking cookie for courses
Previously, progress tracking was handled for lessons and units. This bubbles that information up for courses, thus allowing progress tracking on a per course basis. This wasn't requested, but it seems like a nice-to-have, and isn't difficult to add.

Because the `CourseListViewComponent` currently inherits its logic from the `UnitListViewComponent`, this also means a cookie will be written to the root for the user's status across _all_ topics. That isn't necessary, and we may remove this in the future. That said, it's a single cookie, and affords us the option of reporting whether a user has completed _all_ courses, should we have a need to track that (e.g., for triggering a Google Analytics event).