Present course list as list of panels
This still needs quite a bit of attention. The progress tracking isn't showing up, despite adjusting the selector. The panels are displaying with no padding between them—just as we see on the Unit page—if we don't add a `<p />` between them. The padding is a bit large for a card format. And, finally, they should probably treat the entire panel as clickable.

Ultimately, this may be better served by using the actual `card` format; I'll be revisiting that in a future update. Alternatively, we may want to generalize some of the styles from the home page, which I believe account for some of this.