Removed nested numbering for Lessons
Previously, the lessons used nested numbering—e.g., 1.1, 1.2, 1.3, …—in order to maintain consistency with the Google Courseware app. That was never desirable, however, and was a concession to the limitations of Google's implementation. As such, this functionality is now removed, which conveniently buys us some more space on the left navigation.