Moved MPEG-DASH player configuration to a new `Home.js`
Normally, when an individual view needs its own JavaScript, we create a view-specific file in `/Shared/Scripts/Views`, and call it from the view template. That didn't happen for the homepage, in part because the scripts were so small, but also because there's only one implementation of the page, and thus less caching benefit to sharing the files across template instances.

That said, as the script has grown, it makes more sense to centralize it, not only for consistency, but also to allow it to be subjected to JSLint, Uglify, &c. as part of our build script. This adds an external call, but also reduces the size of the download. And since the video doesn't need to play back by default, the extra HTTP call isn't a big deal. In fact, I've `defer`red the script since it doesn't need to be immediately available on page load.