Moved lazy-loading script to homepage script
Since we now have a `Home.js`, I used it as an opportunity to move the lazy-loading functionality for the images into that script. I'm not a huge fan of using `window.onload` for that, but we can revisit that later.

As part of this, I added more modern comment headers to the JavaScript. This is inconsistent with the other scripts, but more consistent with other Ignia projects. In the future, we'll want to extend this to the rest of the scripts.