Deferred other user interface scripts
As part of the centralization of the MPEG-DASH script (1ec1c5a), I deferred loading of the `Home.js` file. This can also be done for the other user interface oriented scripts, thus helping the main page render faster. That's especially useful for e.g. the Owl Carousel, since that's lower down on the page.